Update to randomic/master

+++ ext/fun.py
@@ -7,24 +7,25 @@ import logging
 from random import randint
 import json
 from os.path import isfile
-import ext.permcheck as permcheck
 
 import discord.ext.commands as commands
+
+import utils.checks as checks
 
 
 def setup(bot):
     'Adds the cog to the provided discord bot'
-    bot.add_cog(Fun(bot))
+    bot.add_cog(Fun(bot, 'memes.json'))
 
 
 class Fun:
     '''A cog defining meme commands'''
-    def __init__(self, bot):
+    def __init__(self, bot, fname):
         self.logger = logging.getLogger(__name__)
         self.bot = bot
         self.guess_number = 0
         self.guess_max = 0
-        self.memelist = ""
+        self.fname = fname
 
     def loadjson(self, jsonname):
         'A function which loads a json, given the filename'
@@ -47,11 +48,12 @@ class Fun:
         except TypeError as error:
             self.logger.error(error)
 
-    def get_status(self):
+    def get_health(self):
         'Returns a string describing the status of this cog'
         response = ""
-        if isfile('memes.json'):
-            response += '\n  \u2714 memes.json exists, memes != dreams'
+        if isfile(self.fname):
+            response += '\n  \u2714 {} exists, memes != dreams'.format(
+                self.fname)
         else:
             response += '\n  \u2716 No meme file found'
         if self.guess_number:
@@ -61,49 +63,48 @@ class Fun:
         return response
 
     @commands.command()
-    @permcheck.check(1)
     async def meme(self, memename: str, imglink: str=""):
         '''Posts a saved imgur link via a specified name
         or saves an imgur link to file under the name.
         Converts all memenames to lower case'''
-        self.memelist = self.loadjson('memes.json')
+        memelist = self.loadjson(self.fname)
         memename = memename.lower()
-        if memename in self.memelist['memes'].keys():
+        if memename in memelist['memes'].keys():
             if imglink != "":
                 await self.bot.say("Meme '%s' already exists!" % memename)
                 self.logger.info("User tried to overwrite a meme")
             else:
-                await self.bot.say(self.memelist['memes'].get(memename))
+                await self.bot.say(memelist['memes'].get(memename))
                 self.logger.info('User posted %s to the chat', memename)
         elif imglink != "":
-            self.memelist['memes'][memename] = imglink
-            self.savejson(self.memelist, 'memes.json')
+            memelist['memes'][memename] = imglink
+            self.savejson(memelist, self.fname)
             await self.bot.say('`{}` added as `{}`!'.format(imglink, memename))
         else:
             await self.bot.say('You entered an invalid meme name')
             self.logger.warning('User entered an invalid meme name')
 
     @commands.command()
-    @permcheck.check(1)
-    async def removememe(self, memename: str):
+    async def removememe(self, memename: str=None):
         '''Removes a meme from file via the specific memename'''
-        self.memelist = self.loadjson('memes.json')
-        memename = memename.lower()
-        if memename in self.memelist['memes'].keys():
-            del self.memelist['memes'][memename]
-            self.logger.info('User removed %s from memes.json', memename)
-            await self.bot.say('%s removed from the memelist.' % memename)
-            self.savejson(self.memelist, 'memes.json')
+        if memename:
+            memename = memename.lower()
+            memelist = self.loadjson(self.fname)
+            if memename in memelist['memes'].keys():
+                del memelist['memes'][memename]
+                self.logger.info('Removed %s from %s', memename, self.fname)
+                await self.bot.say('%s removed from the memelist.' % memename)
+                self.savejson(memelist, self.fname)
+            else:
+                await self.bot.say('No meme named {} exists'.format(memename))
         else:
-            self.logger.warning('User entered invalid memename "%s"', memename)
-            await self.bot.say('You entered an invalid memename.')
+            await self.bot.say('You must enter a memename.')
 
     @commands.command()
-    @permcheck.check(1)
     async def listmemes(self):
         '''Posts a list of the current memes available in the file'''
-        memelist = self.loadjson('memes.json')['memes']
-        response = "**Memes:**\n```"
+        memelist = self.loadjson(self.fname)['memes']
+        response = "**Memes:**\n```\n"
         for key in sorted(memelist.keys()):
             response += "{}:\n  {}\n".format(key, memelist[key])
         response += "```"
@@ -113,6 +114,10 @@ class Fun:
     async def guess(self, ctx, guess: int=0):
         '''Allows the user to guess a number. If there is no number to guess a
         new game is started'''
+        if checks.is_private_channel(ctx):
+            await self.bot.say(
+                'This command cannot be used in a private channel')
+            return
         if self.guess_number:
             if guess < 1 or guess > self.guess_max:
                 await self.bot.delete_message(ctx.message)
@@ -121,13 +126,13 @@ class Fun:
                                    .format(guess, ctx.message.author.mention))
                 self.guess_number = 0
             elif guess < self.guess_number:
+                await self.bot.delete_message(ctx.message)
                 await self.bot.say('{} is too low! Guess again!'
                                    .format(guess))
-                await self.bot.delete_message(ctx.message)
             else:
+                await self.bot.delete_message(ctx.message)
                 await self.bot.say('{} is too high! Guess again!'
                                    .format(guess))
-                await self.bot.delete_message(ctx.message)
         else:
             if guess < 1:
                 guess = 1000
@@ -139,7 +144,6 @@ class Fun:
                                + 'between 1 and {}'.format(guess))
 
     @commands.command()
-    @permcheck.check(2)
     async def emoji(self, emoji: str, number: int=1):
         'Converts a string to :string: aka an emoji'
         emoji = ":" + emoji + ":"
